Local XML Interpreter
This code takes an XML file created by a BLASTn local search and turns
the results into an SQLite database.  Testing still needs to be done to
see if this code can work with other local BLAST searches.
To make this code work, both this program and the XML file need to be on
the desktop.  You can change the parameters if you want to though.
SQLite3 is already installed on versions of python 2.5 or greater, so
nothing new needs to be installed.